+++ src/commonMain/kotlin/collection/NotEmptySet.kt
@@ -12,9 +12,7 @@ import kotlinx.serialization.builtins.SetSerializer
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import kotools.types.ExperimentalSinceKotoolsTypes
 import kotools.types.SinceKotoolsTypes
-import kotools.types.experimental.ExperimentalCollectionApi
 import kotlin.jvm.JvmInline
 
 /**
@@ -72,41 +70,6 @@ public fun <E> Collection<E>.toNotEmptySet(): Result<NotEmptySet<E>> =
         val elements: Set<E> = toSet()
         NotEmptySet(elements)
     }
-
-/**
- * Returns a [NotEmptySet] containing all the elements of this collection, or
- * throws an [IllegalArgumentException] if this collection is
- * [empty][Collection.isEmpty].
- *
- * ```kotlin
- * var collection: Collection<Int> = setOf(1, 2, 3, 1)
- * var result: NotEmptySet<Int> = collection.toNotEmptySetOrThrow()
- * println(result) // [1, 2, 3]
- *
- * collection = emptySet()
- * collection.toNotEmptySetOrThrow() // IllegalArgumentException
- * ```
- *
- * Please note that changes made to the original collection will not be
- * reflected on the resulting [NotEmptyList].
- *
- * ```kotlin
- * val original: MutableCollection<Int> = mutableSetOf(1, 2, 3, 1)
- * val notEmptySet: NotEmptySet<Int> = original.toNotEmptySetOrThrow()
- * println(original) // [1, 2, 3]
- * println(notEmptySet) // [1, 2, 3]
- *
- * original.clear()
- * println(original) // []
- * println(notEmptySet) // [1, 2, 3]
- * ```
- */
-@ExperimentalCollectionApi
-@ExperimentalSinceKotoolsTypes("4.3.1")
-public fun <E> Collection<E>.toNotEmptySetOrThrow(): NotEmptySet<E> {
-    val elements: Set<E> = toSet()
-    return NotEmptySet(elements)
-}
 
 /**
  * Represents a set with at least one element of type [E].