+++ CHANGELOG.md
@@ -78,7 +78,7 @@ All notable changes to this project will be documented in this file.
   - `toStrictlyPositiveDoubleOrNull` and `toStrictlyPositiveDoubleOrThrow`
   - `toNotBlankStringOrNull` and `toNotBlankStringOrThrow`
   - `toNotEmptyListOrNull` and `toNotEmptyListOrThrow`
-  - `toNotEmptySetOrNull`.
+  - `toNotEmptySetOrNull` and `toNotEmptySetOrThrow`.
 
 [#258]: https://github.com/kotools/types/issues/258